Split integration suite by resource area; fix a real refresh-token bug
Split integration/uaa_test.go into one file per resource area (clients,
users, groups, tokens, misc) plus a shared suite_test.go, each its own
top-level Describe instead of one giant Ordered block. Ginkgo's Ordered
containers skip every remaining spec once one fails, so a single failure
anywhere was hiding the pass/fail status of everything unrelated to it.
Verified against a live UAA: with the two known-bad specs left pending,
splitting drops "22 skipped" to "0 skipped" for a real failure.

While isolating those two pending specs against a freshly-booted UAA
(confirming they're not local-state artifacts), found and fixed a real
uaa-cli bug: RefreshTokenValidations rejected an explicitly-empty
client_secret as "missing", breaking refresh-token for public clients
(e.g. UAA's built-in "cf" client) entirely. get-password-token already
allows an empty secret for the same reason; refresh-token now matches.

The other pending spec (unlock-user) hits a real, ~9-year-old bug in
cloudfoundry/uaa itself, not uaa-cli: ScimUserEndpoints.updateAccountStatus
is missing @ResponseBody (dropped in commit 6159e2f4, 2016), so
PATCH /Users/{userId}/status always 500s. Left pending with a comment
until that's fixed upstream.

Also fixed the "clients" area's own test fixture: the throwaway client
needs the clients.secret authority to change its own secret via
change-client-secret, which it didn't have.
